+++ src/Modules/Basket/BasketGateway.php
@@ -285,7 +285,7 @@ class BasketGateway extends BaseGateway
 				fs.name AS fs_name,
 				fs.photo AS fs_photo,
 				fs.id AS fs_id,
-				fs.sleep_status AS fs_sleep_status,
+				fs.is_sleeping AS fs_is_sleeping,
 				b.id,
 				b.description
 			FROM fs_basket_anfrage a
@@ -343,7 +343,7 @@ class BasketGateway extends BaseGateway
 				fs.id AS fs_id,
 				fs.name AS fs_name,
 				fs.photo AS fs_photo,
-				fs.sleep_status AS fs_sleep_status
+				fs.is_sleeping AS fs_is_sleeping
 			FROM fs_basket b
             JOIN fs_foodsaver fs ON b.foodsaver_id = fs.id
             WHERE