+++ CHANGELOG.md
@@ -92,6 +92,7 @@
 - Clean up code for maintenance and background tasks !3735 @\_fridtjof_
 - Port the dashboard controller to Symfony !3736 @\_fridtjof_
 - Entries from the content API are rendered in Vue !3743 @alex.simm
+- Remove `FoodsaverForAvatar` in favor of `Profile` #1998 !3713 @AntonBallmaier
 
 ## Dev/Test/CI stuff
 - Improved API tests runtime !3506 @alex.simm